notice search query edit

+++ src/main/java/com/jinjin/jintranet/notice/dto/NoticeSearchDTO.java
@@ -1,14 +1,13 @@
 package com.jinjin.jintranet.notice.dto;
 
-import java.time.LocalDateTime;
-
 import com.jinjin.jintranet.model.Member;
 import com.jinjin.jintranet.model.Notice;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.time.LocalDateTime;
 
 @Getter
 @Setter
@@ -24,13 +23,13 @@ public class NoticeSearchDTO {
 	
 	private LocalDateTime crtDt;
 	
-	private Integer attachesCnt;
+	private Long attachesCnt;
 	
 	public NoticeSearchDTO(Notice notice) {
 		this.id = notice.getId();
 		this.title = notice.getTitle();
 		this.member = notice.getMember();
 		this.crtDt = notice.getCrtDt();
-		this.attachesCnt = (int) notice.getAttaches().stream().filter(a -> a.getDeletedBy() ==null).count() ;
+		this.attachesCnt = notice.getAttaches().stream().filter(a -> a.getDeletedBy() ==null).count() ;
 	}
 }
